# packages/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TOMA API",
    "version": "1.0.0",
    "description": "Training management system API. Contract generated from @toma/shared zod schemas."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "default": "about:blank"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/courses/{id}/attendance-grid": {
      "get": {
        "tags": [
          "attendance"
        ],
        "summary": "The per-day attendance grid HR fills in at the end of each course day (requirement #9)",
        "parameters": [
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ]
            },
            "required": false,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Attendance grid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "courseId": {
                      "type": "integer"
                    },
                    "courseTitle": {
                      "type": "string"
                    },
                    "sessions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "startsAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "endsAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "startsAt",
                          "endsAt"
                        ]
                      }
                    },
                    "rows": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "employeeId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "employeeName": {
                            "type": "string"
                          },
                          "present": {
                            "type": "array",
                            "items": {
                              "type": "boolean"
                            }
                          }
                        },
                        "required": [
                          "employeeId",
                          "employeeName",
                          "present"
                        ]
                      }
                    }
                  },
                  "required": [
                    "courseId",
                    "courseTitle",
                    "sessions",
                    "rows"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/courses/{id}/attendance": {
      "put": {
        "tags": [
          "attendance"
        ],
        "summary": "HR marks one person present/absent for one day; an absence opens a justification (#9)",
        "parameters": [
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ]
            },
            "required": false,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "employeeId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "sessionStart": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "present": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "employeeId",
                  "sessionStart",
                  "present"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Recorded"
          },
          "403": {
            "description": "Not permitted for this role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/justifications/{id}/submit": {
      "post": {
        "tags": [
          "attendance"
        ],
        "summary": "Submit a reason for a no-show (employee or their manager)",
        "parameters": [
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ]
            },
            "required": false,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "reason"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated justification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "exclusiveMinimum": 0
                    },
                    "courseId": {
                      "type": "integer",
                      "exclusiveMinimum": 0
                    },
                    "courseTitle": {
                      "type": "string"
                    },
                    "employeeId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "employeeName": {
                      "type": "string"
                    },
                    "sessionDate": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "reason": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "requested",
                        "submitted",
                        "accepted",
                        "rejected"
                      ]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "courseId",
                    "courseTitle",
                    "employeeId",
                    "employeeName",
                    "sessionDate",
                    "reason",
                    "status",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/justifications/{id}/review": {
      "post": {
        "tags": [
          "attendance"
        ],
        "summary": "HR accepts or rejects a submitted justification",
        "parameters": [
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ]
            },
            "required": false,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "decision": {
                    "type": "string",
                    "enum": [
                      "accept",
                      "reject"
                    ]
                  }
                },
                "required": [
                  "decision"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reviewed justification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "exclusiveMinimum": 0
                    },
                    "courseId": {
                      "type": "integer",
                      "exclusiveMinimum": 0
                    },
                    "courseTitle": {
                      "type": "string"
                    },
                    "employeeId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "employeeName": {
                      "type": "string"
                    },
                    "sessionDate": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "reason": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "requested",
                        "submitted",
                        "accepted",
                        "rejected"
                      ]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "courseId",
                    "courseTitle",
                    "employeeId",
                    "employeeName",
                    "sessionDate",
                    "reason",
                    "status",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
...
}